Add files via upload
First run the server so it will be ready to acknowledge client and then run the UDPclient to program to establish communication between them. Since it runs on the localhost please enter your ip address in order to run the files. The local host is set to default 12345. You will be good to go from there!